move jwtCheck to middleware/authorise
this change is needed so that we can secure routes with token checking
middleware
