+++ index.js
@@ -4,5 +4,9 @@ const port = process.env.PORT || 9000;
 
 server.listen(port, () => {
   // eslint-disable-next-line no-console
-  console.log('=== Listening on port %d in %s mode ===', port, server.settings.env);
+  console.log(
+    '=== Listening on port %d in %s mode ===',
+    port,
+    server.settings.env
+  );
 });
